pyproject.toml
[tool.poetry]
name = "pyfuncify"
version = "0.1.3"
description = ""
authors = ["Col Perks <wild.fauve@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyMonad = "^2.4.0"
python-dateutil = "^2.8.2"
jwcrypto = "^1.0"
PyJWT = "^2.1.0"
pino = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
